Try to maintain backward compatibility of APIs
Instead of extracting an interface from QueryRewriteContext, we
can keep it as a base class and subclass it for QueryCoordinatorContext.
