+++ server/src/main/java/org/opensearch/index/query/QueryCoordinatorContext.java
@@ -8,16 +8,11 @@
 
 package org.opensearch.index.query;
 
-import org.opensearch.client.Client;
 import org.opensearch.common.annotation.PublicApi;
-import org.opensearch.core.action.ActionListener;
-import org.opensearch.core.common.io.stream.NamedWriteableRegistry;
-import org.opensearch.core.xcontent.NamedXContentRegistry;
 import org.opensearch.search.pipeline.PipelinedRequest;
 
 import java.util.HashMap;
 import java.util.Map;
-import java.util.function.BiConsumer;
 
 /**
  * The QueryCoordinatorContext class implements the QueryRewriteContext interface and provides
@@ -29,53 +24,12 @@ import java.util.function.BiConsumer;
  * @since 2.19.0
  */
 @PublicApi(since = "2.19.0")
-public class QueryCoordinatorContext implements QueryRewriteContext {
-    private final QueryRewriteContext rewriteContext;
+public class QueryCoordinatorContext extends QueryRewriteContext {
     private final PipelinedRequest searchRequest;
 
-    public QueryCoordinatorContext(QueryRewriteContext rewriteContext, PipelinedRequest searchRequest) {
-        this.rewriteContext = rewriteContext;
-        this.searchRequest = searchRequest;
-    }
-
-    @Override
-    public NamedXContentRegistry getXContentRegistry() {
-        return rewriteContext.getXContentRegistry();
-    }
-
-    @Override
-    public long nowInMillis() {
-        return rewriteContext.nowInMillis();
-    }
-
-    @Override
-    public NamedWriteableRegistry getWriteableRegistry() {
-        return rewriteContext.getWriteableRegistry();
-    }
-
-    @Override
-    public QueryShardContext convertToShardContext() {
-        return rewriteContext.convertToShardContext();
-    }
-
-    @Override
-    public void registerAsyncAction(BiConsumer<Client, ActionListener<?>> asyncAction) {
-        rewriteContext.registerAsyncAction(asyncAction);
-    }
-
-    @Override
-    public boolean hasAsyncActions() {
-        return rewriteContext.hasAsyncActions();
-    }
-
-    @Override
-    public void executeAsyncActions(ActionListener listener) {
-        rewriteContext.executeAsyncActions(listener);
-    }
-
-    @Override
-    public boolean validate() {
-        return rewriteContext.validate();
+    public QueryCoordinatorContext(QueryRewriteContext parent, PipelinedRequest pipelinedRequest) {
+        super(parent.getXContentRegistry(), parent.getWriteableRegistry(), parent.client, parent.nowInMillis, parent.validate());
+        this.searchRequest = pipelinedRequest;
     }
 
     @Override
