+++ server/src/test/java/org/opensearch/index/query/RewriteableTests.java
@@ -45,7 +45,7 @@ import java.util.function.Supplier;
 public class RewriteableTests extends OpenSearchTestCase {
 
     public void testRewrite() throws IOException {
-        QueryRewriteContext context = new BaseQueryRewriteContext(null, null, null, null);
+        QueryRewriteContext context = new QueryRewriteContext(null, null, null, null);
         TestRewriteable rewrite = Rewriteable.rewrite(
             new TestRewriteable(randomIntBetween(0, Rewriteable.MAX_REWRITE_ROUNDS)),
             context,
@@ -65,7 +65,7 @@ public class RewriteableTests extends OpenSearchTestCase {
     }
 
     public void testRewriteAndFetch() throws ExecutionException, InterruptedException {
-        BaseQueryRewriteContext context = new BaseQueryRewriteContext(null, null, null, null);
+        QueryRewriteContext context = new QueryRewriteContext(null, null, null, null);
         PlainActionFuture<TestRewriteable> future = new PlainActionFuture<>();
         Rewriteable.rewriteAndFetch(new TestRewriteable(randomIntBetween(0, Rewriteable.MAX_REWRITE_ROUNDS), true), context, future);
         TestRewriteable rewrite = future.get();
@@ -83,7 +83,7 @@ public class RewriteableTests extends OpenSearchTestCase {
     }
 
     public void testRewriteList() throws IOException {
-        BaseQueryRewriteContext context = new BaseQueryRewriteContext(null, null, null, null);
+        QueryRewriteContext context = new QueryRewriteContext(null, null, null, null);
         List<TestRewriteable> rewriteableList = new ArrayList<>();
         int numInstances = randomIntBetween(1, 10);
         rewriteableList.add(new TestRewriteable(randomIntBetween(1, Rewriteable.MAX_REWRITE_ROUNDS)));
